requirements.txt
beautifulsoup4==4.11.1
black==22.8.0
blessed==1.19.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
idna==3.4
inquirer==2.10.0
lxml==4.9.1
mypy-extensions==0.4.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pycodestyle==2.9.1
python-editor==1.0.4
readchar==4.0.3
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
toml==0.10.2
tomli==2.0.1
urllib3==1.26.12
wcwidth==0.2.5
